requirements.txt
firebase-admin==6.4.0
python-dotenv==1.0.1
langdetect==1.0.9
deep-translator==1.11.4
groq==0.24.0